fix: trigger the frontend_ready event
